fix(copy-trading): add ::numeric casts to FreezeFromBucket/UnfreezeBucket
The WHERE clause `available_capital >= $2 + $3` (and `frozen_capital >= $2`
is fine but `$2 + $3` anywhere was not) caused every copy-trade open to
fail with `operator is not unique: unknown + unknown (SQLSTATE 42725)`.

Root cause: pgx binds float64 params without pinning a Postgres type.
When two `unknown`-typed params are added together without a column
anchor on either side (as in a WHERE `>=` comparison), Postgres cannot
resolve the `+` operator. SET clauses with a column on the left
(`available_capital + $2 + $3 - $4`) still work via left-to-right type
inference, but the WHERE side exploded.

Impact: since commit 7d5f9b2 (2026-04-15), 100% of copy-trade opens
failed silently in the skip path, leaving followers watching their
trader trade with no mirror fills. Uncovered while investigating why
arron's recent opens did not cascade to aliex.

Fix: add explicit `::numeric` casts to every parameter in both
FreezeFromBucket and UnfreezeBucket arithmetic — WHERE and SET — and
leave a comment so future refactors don't reintroduce the bug. Other
similar patterns (SettleToBucket, wallet SettleTrade) are empirically
safe because they are column-anchored in SET with no naked `$N + $M`
in WHERE, but this fix focuses on the confirmed failure path.

+++ tongxin-go/internal/repository/trader_repo.go
@@ -578,12 +578,19 @@ func (r *TraderRepo) FreezeFromBucket(ctx context.Context, copyTradingID string,
 	if fee < 0 {
 		return fmt.Errorf("fee must be >= 0: %v", fee)
 	}
+	// NOTE: `$2::numeric + $3::numeric` 是必要的显式类型转换。
+	// pgx 绑定 float64 参数时不会固定 parameter 类型，若写成 `$2 + $3`，
+	// PostgreSQL 在没有列上下文（比如 WHERE 里的 `>= $2 + $3`）的情况下
+	// 会把两个参数都当 unknown，报 `operator is not unique: unknown + unknown
+	// (SQLSTATE 42725)`。SET 子句里虽然 `available_capital - $2 - $3` 能借
+	// 列类型推断，但为了统一和防御，一律加 ::numeric。
 	tag, err := r.pool.Exec(ctx, `
 		UPDATE copy_trading
-		SET available_capital = available_capital - $2 - $3,
-		    frozen_capital    = frozen_capital + $2,
+		SET available_capital = available_capital - $2::numeric - $3::numeric,
+		    frozen_capital    = frozen_capital + $2::numeric,
 		    updated_at = NOW()
-		WHERE id = $1 AND status = 'active' AND available_capital >= $2 + $3
+		WHERE id = $1 AND status = 'active'
+		  AND available_capital >= $2::numeric + $3::numeric
 	`, copyTradingID, margin, fee)
 	if err != nil {
 		return fmt.Errorf("freeze from bucket: %w", err)
@@ -601,12 +608,14 @@ func (r *TraderRepo) UnfreezeBucket(ctx context.Context, copyTradingID string, m
 	if margin <= 0 {
 		return nil
 	}
+	// 见 FreezeFromBucket 的说明：$2 / $3 必须加 ::numeric 显式类型，
+	// 否则 `$2 + $3` 在无列上下文时会被 pgx 当 unknown 类型报错。
 	tag, err := r.pool.Exec(ctx, `
 		UPDATE copy_trading
-		SET frozen_capital    = frozen_capital - $2,
-		    available_capital = available_capital + $2 + $3,
+		SET frozen_capital    = frozen_capital - $2::numeric,
+		    available_capital = available_capital + $2::numeric + $3::numeric,
 		    updated_at = NOW()
-		WHERE id = $1 AND frozen_capital >= $2
+		WHERE id = $1 AND frozen_capital >= $2::numeric
 	`, copyTradingID, margin, fee)
 	if err != nil {
 		return fmt.Errorf("unfreeze bucket: %w", err)
